drivers: dcsub: Update the XDcSub_SetNonLiveLatency API
Update the datatype of NonLiveLatency variable and set accurate value of
nonlive latency in the XDcSub_SetNonLiveLatency API.


Acked-for-series: Varunkumar Allagadapa <varunkumar.allagadapa@amd.com>

+++ XilinxProcessorIPLib/drivers/dcsub/src/xdc.h
@@ -422,7 +422,7 @@ typedef struct {
 
 	u32 VidFrameSwitchCtrl;
 
-	u16 NonLiveLatency;
+	u32 NonLiveLatency;
 
 	u32 StcEn;
 	u32 StcInitVal;
